+++ src/md5.h
@@ -70,7 +70,7 @@ typedef struct {
 void MD5_Transform (UINT4* buf, UINT4* in);
 
 void MD5Init(MD5_CTX* mdContext, unsigned long pseudoRandomNumber = 0);
-void MD5Update(MD5_CTX* mdContext, const unsigned char* inBuf, unsigned int inLen);
+void MD5Update(MD5_CTX* mdContext, const unsigned char* inBuf, size_t inLen);
 void MD5Final(MD5_CTX* mdContext);
 
 #endif /* ___MD5_H___ included */
